+++ src/app/locales/en/components/qos-modal/qos-entry-charts.js
@@ -1,0 +1,17 @@
+export default {
+  header: 'Transfer statistics',
+  unknownStorage: 'Storage#{{id}}',
+  unknownProvider: 'Provider#{{id}}',
+  titles: {
+    inbound: 'Inbound',
+    outbound: 'Outbound',
+  },
+  axes: {
+    bytes: 'Bytes',
+    files: 'Files',
+  },
+  series: {
+    totalBytes: 'Total',
+    totalFiles: 'Files',
+  },
+};